Crate imag_documentation

imag

This is the developer documentation for the imag personal information management suite for the commandline.

For the user documentation, have a look at the 'doc' subtree in the repository which can be compiled to HTML or PDF using pandoc (and might be a bit outdated as imag is not yet released for use).

General

Some things from the user documentation might be helpful for developers as well, so make sure to at least skim over it if you want to contribute to the imag source.

Also make sure you had a look at the CONTRIBUTING file and the developers certificate of origin, which we also have in the CONTRIBUTING file, by the way.

Contributing

All things relevant for contributing are descripbed in the CONTRIBUTING file, but here are some additional notes:

We have a editorconfig file in the repository. Would be nice if you'd use it

We have default.nix file, which can be used to install dependencies in nix-shell environments. If you have a installation of the nix package manager, feel free to use this opportunity to be pure.

If you want to play around with imag, use the imagrc.toml file from the repository, we try to keep it up to date.

You can use the Makefile to build things (if you have all dependencies and cargo/rustc installed, of course).

It is a real advantage to use cargo-check when developing things - it speeds you up, beleive me!